- add toolbar in each fragment instead of in the main activity to implement the search function in each fragment - changed colors for stars in ratingBar - implemented onBackPressed method in main Activity - changed design of the edit text of the search function (remove the underline and add a cross to clear the text)

+++ app/src/main/java/com/example/blanche/go4lunch/models/RestaurantPlace.java
@@ -9,7 +9,7 @@ import java.util.Map;
 public class RestaurantPlace {
 
     int maxLike;
-    int like;
+    float like;
     String uid;
     List<String> usersWhoLiked;
     List<Map<String, Integer> > usersWhoLiked2;
@@ -29,7 +29,7 @@ public class RestaurantPlace {
         this.like = like;
     }
 
-    public int getLike() {
+    public float getLike() {
         return like;
     }
